+++ packages/networks/src/index.spec.ts
@@ -67,6 +67,26 @@ describe('availableNetworks', (): void => {
       )
     ).toEqual([]);
   });
+
+  it('has all the correct fields', (): void => {
+    expect(availableNetworks[0]).toEqual({
+      decimals: [10],
+      displayName: 'Polkadot Relay Chain',
+      genesisHash: [
+        '0x91b171bb158e2d3848fa23a9f1c25182fb8e20313b2c1eb49219da7a70ce90c3'
+      ],
+      hasLedgerSupport: true,
+      icon: 'polkadot',
+      isIgnored: false,
+      isTestnet: false,
+      network: 'polkadot',
+      prefix: 0,
+      slip44: 354,
+      standardAccount: '*25519',
+      symbols: ['DOT'],
+      website: 'https://polkadot.network'
+    });
+  });
 });
 
 describe('allNetworks', (): void => {